more testing issues

+++ install_scripts/startService.py
@@ -34,23 +34,12 @@ Group=root
 WantedBy=multi-user.target
 """
 
-def set_permissions():
-    uid = pwd.getpwnam("root").pw_uid
-    gid = grp.getgrnam("root").gr_gid
-    os.chown(working_directory, uid, gid)
-    for root, dirs, files in os.walk(working_directory):
-        for dir_ in dirs:
-            os.chown(os.path.join(root, dir_), uid, gid)
-        for file_ in files:
-            # Skip files created by build_web.py
-            if 'build_web.py' not in file_:
-                os.chown(os.path.join(root, file_), uid, gid)
 
 def check_index_file():
     index_file_path = os.path.join(working_directory, 'templates/index.htm')
     if not os.path.exists(index_file_path):
         print(f"Error: {index_file_path} not found. Running build_web.py...")
-        build_web_script = os.path.join(working_directory, 'build_web.py')
+        build_web_script = os.path.join(working_directory, 'install_scripts/build_web.py')
         # Get the owner of the working directory
         dir_owner = pwd.getpwuid(os.stat(working_directory).st_uid).pw_name
         # Run build_web.py as the directory owner
@@ -63,11 +52,21 @@ def check_index_file():
 
 def install_requirements():
     requirements_file = os.path.join(working_directory, 'requirements.txt')
+    marker_file = os.path.join(working_directory, '.requirements_installed')
+
+    # Check if the marker file exists
+    if os.path.exists(marker_file):
+        print("Requirements already installed. Skipping installation.")
+        return
+
     if os.path.exists(requirements_file):
         print("Installing packages from requirements.txt...")
         if os.system(f'python3 -m pip install --break-system-packages -r {requirements_file}') != 0:
             print("Error: Failed to install required packages.")
             sys.exit(1)
+        # Create the marker file to indicate successful installation
+        with open(marker_file, 'w') as marker:
+            marker.write("Requirements installed.")
     else:
         print("requirements.txt not found. Skipping package installation.")
 
@@ -82,8 +81,6 @@ def create_service():
         shutil.copy(local_service_file, service_file_path)
         print("Systemd service file created/updated successfully.")
 
-    # Set the correct permissions for the working directory
-    #set_permissions()
 
     # Reload systemd, enable and start the service
     os.system('systemctl daemon-reload')
